+++ src/headless/headless.js
@@ -6,11 +6,10 @@ import "./plugins/adhoc.js";            // XEP-0050 Ad Hoc Commands
 import "./plugins/bookmarks/index.js";  // XEP-0199 XMPP Ping
 import "./plugins/bosh.js";             // XEP-0206 BOSH
 import "./plugins/caps/index.js";       // XEP-0115 Entity Capabilities
-import "./plugins/carbons.js";          // XEP-0280 Message Carbons
 import "./plugins/chat/index.js";       // RFC-6121 Instant messaging
 import "./plugins/chatboxes/index.js";
 import "./plugins/disco/index.js";      // XEP-0030 Service discovery
-import "./plugins/headlines.js";        // Support for headline messages
+import "./plugins/headlines/index.js";  // Support for headline messages
 import "./plugins/mam/index.js";        // XEP-0313 Message Archive Management
 import "./plugins/muc/index.js";        // XEP-0045 Multi-user chat
 import "./plugins/ping/index.js";       // XEP-0199 XMPP Ping