+++ src/shared/chat/message-actions.js
@@ -65,16 +65,15 @@ class MessageActions extends CustomElement {
         `;
     }
 
-    onMessageEditButtonClicked (ev) {
+    async onMessageEditButtonClicked (ev) {
         ev.preventDefault();
         const currently_correcting = this.model.collection.findWhere('correcting');
         // TODO: Use state intead of DOM querying
         // Then this code can also be put on the model
         const unsent_text = u.ancestor(this, '.chatbox')?.querySelector('.chat-textarea')?.value;
         if (unsent_text && (!currently_correcting || currently_correcting.getMessageText() !== unsent_text)) {
-            if (!confirm(__('You have an unsent message which will be lost if you continue. Are you sure?'))) {
-                return;
-            }
+            const result = await api.confirm(__('You have an unsent message which will be lost if you continue. Are you sure?'));
+            if (!result) return;
         }
         if (currently_correcting !== this.model) {
             currently_correcting?.save('correcting', false);