+++ webpack/webpack.headless.js
@@ -4,13 +4,15 @@ const path = require('path');
 const { merge } = require("webpack-merge");
 
 module.exports = merge(common, {
-    entry: "@converse/headless/headless.js",
+    entry: {
+        "converse-headless": "@converse/headless/headless.js",
+        "converse-headless.min": "@converse/headless/headless.js",
+    },
     output: {
         path: path.resolve(__dirname, '../src/headless/dist'), // Output path for generated bundles
-        filename: 'converse-headless.min.js',
+        filename: "[name].js",
         chunkFilename: '[name].js'
     },
     mode: "production",
-    devtool: "source-map",
 });